Unit7/face_recognition_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
face_recognition
mysql-connector-python
